Move tests from ControlRequestTest to dedicated test classes and simplify a little
